fix: cpu and cuda switch support
