+++ .gitignore
@@ -73,7 +73,6 @@ cover/
 # Environments
 .env
 .venv
-env
 venv
 ENV
 env.bak/
