+++ dreamer_node/agents/dreamer_agent.py
@@ -12,12 +12,30 @@ import pathlib
 import torch
 from torch import distributions as torchd
 
-from util.constants import Constants
-from util.constants import Config
+from config import Config
 
 from dreamer.dream import Dreamer
 from dreamer.dream import Parallel, Damy
 import dreamer.tools as tools
+
+
+class Constants:
+    """
+    Class to hold constants for the car.
+    """
+
+    def __init__(self):
+        self.FORWARD_SCAN_ARC = (np.deg2rad(-90.0), np.deg2rad(+90.0))
+        # Topics
+        self.LIDAR_TOPIC = "/scan"
+        self.DRIVE_TOPIC = "/drive"
+        self.ODOMETRY_TOPIC = "/ego_racecar/odom"
+
+        # action space
+        self.min_steering = -0.418
+        self.max_steering = 0.418
+        self.min_speed = 1.5
+        self.max_speed = 19.67
 
 
 class DreamerRacer(Node):
