+++ dreamer_node/tests/test_racerenv.py
@@ -1,7 +1,7 @@
 import unittest
 
 from env.racerenv import RacerEnv
-from util.constants import Constants
+from config import Config
 
 
 class TestRacerEnv(unittest.TestCase):
@@ -48,23 +48,6 @@ class TestRacerEnv(unittest.TestCase):
         )
         self.assertEqual(
             environment.denormalize(0.5, -60, 20), 0, "Should be 75th percentile."
-        )
-
-    def test_constants(self):
-        environment = RacerEnv()
-        constants = Constants()
-
-        self.assertEqual(
-            environment._min_steering, constants.min_steering, "Should be equal."
-        )
-        self.assertEqual(
-            environment._max_steering, constants.max_steering, "Should be equal."
-        )
-        self.assertEqual(
-            environment._min_speed, constants.min_speed, "Should be equal."
-        )
-        self.assertEqual(
-            environment._max_speed, constants.max_speed, "Should be equal."
         )
